requirements.txt
google-generativeai
pinecone
sentence-transformers
flask
flask-cors
streamlit
python-dotenv
